rework clerk example
